requirements.txt
asgiref==3.5.0
autopep8==1.6.0
backports.zoneinfo==0.2.1
click==8.0.1
colorama==0.4.4
distlib==0.3.4
Django==4.0.4
django-autoslug==1.9.8
filelock==3.6.0
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
Pillow==9.1.0
platformdirs==2.5.2
protobuf==4.21.2
psycopg2-binary==2.9.3
pycodestyle==2.8.0
six==1.16.0
SQLAlchemy==1.4.17
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
virtualenv==20.14.1
Werkzeug==2.0.1
